refactor: canvas area
-Move canvas area logic to it's own UI component.
