+++ client/src/components/utils/Types.tsx
@@ -1,0 +1,17 @@
+interface CanvasAreaProps {
+  devices: Device[];
+  selectedDeviceId: string | null;
+  onCanvasClick?: (x: number, y: number) => void;
+  viewport: { x: number; y: number };
+  setViewport: React.Dispatch<React.SetStateAction<{ x: number; y: number }>>;
+}
+
+type Device = {
+  id: string;
+  x: number;
+  y: number;
+  type: string;
+  state: Record<string, any>;
+};
+
+export type { CanvasAreaProps, Device };
